Create commands.h
Of course, most of this file was just ideas of mine. I have never used GitHub before, so I am not sure if this is going to immediately be added or if you check it. I assume it's the latter because it says "Propose new file" under this edit box. 

Anyway, basic summary:
WriteCommand() get's the user input, then ProcessCommand() processes it. The overloaded ProcessCommand(int) is a predetermined command in the code. Say you want a specific command called, you would call this. Now, I am contemplating on whether or not we implement functions for each error. I may just do one function, get the error ID, then throw an exception or something.
How's it sound? Really glad to have finally done something with this...